ft_cha on the way

+++ include/ft_printf.h
@@ -23,11 +23,11 @@ size_t	ft_strlcpy(char *dst, const char *src, size_t size);
 void	*ft_memcpy(void *restrict dest, const void *restrict src, size_t n);
 
 char	*ft_str(char *fptr, long int flen, va_list *ap);
-char	*ft_cha(char *fptr, int	flen, va_list *ap);
-char	*ft_ptr(char *fptr, int	flen, va_list *ap);
-char	*ft_int(char *fptr, int	flen, va_list *ap);
-char	*ft_hex(char *fptr, int	flen, va_list *ap);
-char	*ft_HEX(char *fptr, int	flen, va_list *ap);
-char	*ft_uns(char *fptr, int	flen, va_list *ap);
+char	*ft_cha(char *fptr, long int flen, va_list *ap);
+char	*ft_ptr(char *fptr, long int flen, va_list *ap);
+char	*ft_int(char *fptr, long int flen, va_list *ap);
+char	*ft_hex(char *fptr, long int flen, va_list *ap);
+char	*ft_HEX(char *fptr, long int flen, va_list *ap);
+char	*ft_uns(char *fptr, long int flen, va_list *ap);
 #endif
